+++ src/inspect_ai/analysis/beta/_dataframe/messages/table.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Callable, Literal, TypeAlias
 
+from inspect_ai.log._file import list_eval_logs
 from inspect_ai.model._chat_message import ChatMessage
 
 if TYPE_CHECKING:
@@ -22,43 +23,36 @@ MessageFilter: TypeAlias = (
 
 @overload
 def messages_df(
-    logs: LogPaths,
+    logs: LogPaths = list_eval_logs(),
     columns: list[Column] = MessageColumns,
     filter: MessageFilter | None = None,
-    recursive: bool = True,
-    reverse: bool = False,
     strict: Literal[True] = True,
 ) -> "pd.DataFrame": ...
 
 
 @overload
 def messages_df(
-    logs: LogPaths,
+    logs: LogPaths = list_eval_logs(),
     columns: list[Column] = MessageColumns,
     filter: MessageFilter | None = None,
-    recursive: bool = True,
-    reverse: bool = False,
     strict: Literal[False] = False,
 ) -> tuple["pd.DataFrame", ColumnErrors]: ...
 
 
 def messages_df(
-    logs: LogPaths,
+    logs: LogPaths = list_eval_logs(),
     columns: list[Column] = MessageColumns,
     filter: MessageFilter | None = None,
-    recursive: bool = True,
-    reverse: bool = False,
     strict: bool = True,
 ) -> "pd.DataFrame" | tuple["pd.DataFrame", ColumnErrors]:
     """Read a dataframe containing messages from a set of evals.
 
     Args:
        logs: One or more paths to log files or log directories.
+          Defaults to the contents of the currently active log directory
+          (e.g. ./logs or INSPECT_LOG_DIR).
        columns: Specification for what columns to read from log files.
        filter: List of message role types to include or callable that performs the filter.
-       recursive: Include recursive contents of directories (defaults to `True`)
-       reverse: Reverse the order of the dataframe (by default, items
-          are ordered from oldest to newest).
        strict: Raise import errors immediately. Defaults to `True`.
           If `False` then a tuple of `DataFrame` and errors is returned.
 
@@ -80,8 +74,6 @@ def messages_df(
     return _read_samples_df(
         logs=logs,
         columns=columns,
-        recursive=recursive,
-        reverse=reverse,
         strict=strict,
         detail=detail,
     )